The great migration. All BMP3xx C code was moved into the renamed BMP388 C++ Library. The bmp3_dev struct is no longer needed, meaning the class can be used for multiple sensors. Tasks updated to call the new class. Builds, but not tested.
