+++ src/SensorTask.hpp
@@ -9,9 +9,9 @@ class SensorTask;
 #include "support/driver_init.h"
 #include "printf.h"
 #include "support/atmel_start_pins.h"
-#include "ADXL375.hpp"
-#include "BMP3xx.hpp"
-#include "BMI088.hpp"
+#include "periph/ADXL375/ADXL375.hpp"
+#include "periph/BMP388/BMP388.hpp"
+#include "periph/BMI088/BMI088.hpp"
 #include "main.hpp"
 #include "ArduinoJson.h"
 #include "AltFilterTask.hpp"
